I2C Work
Attempting to get I2C to work. Not quite there. Likely an issue with addressing the SDA and SCL pins to their output. I have gotten GPIO to work referencing the "GXX" version of the pins. Need to check the logic output to see if the SDA and SCL pins are firing on a i2c.scan() call.